requirements.txt
alembic==1.13.2
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
banal==1.0.6
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colab-ssh==0.3.27
colorama==0.4.6
comm==0.2.2
cryptography==42.0.8
dataset==1.6.2
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.0.1
django-allauth==0.61.1
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
docker==7.1.0
drf-yasg==1.21.7
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
fqdn==1.5.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
install==1.3.5
ipykernel==6.29.4
ipython==8.25.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-http-over-ws==0.0.8
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.1
notebook_shim==0.2.4
oauthlib==3.2.2
overrides==7.7.0
packaging==23.2
pandocfilters==1.5.1
parso==0.8.4
platformdirs==4.2.2
portalocker==2.8.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2023.4
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==1.4.52
sqlparse==0.4.4
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
